+++ crates/api-desc/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Minor
 
+- Add `scheduling::abort()`
 - Require the `U32` type to implement `bytemuck::Pod`.
 - Add `debug::time()`, `debug::perf()`, and `debug::Perf` for simple performance
   measurement